refactor: extract usePayloadSession hook from App.tsx
