+++ src/App.tsx
@@ -5,7 +5,6 @@ import {
   useEffect,
   useEffectEvent,
   useMemo,
-  useRef,
   useState,
 } from "react";
 import { getVersion } from "@tauri-apps/api/app";
@@ -66,19 +65,12 @@ import {
 import type { ToolbarItem } from "./components/toolbar/types";
 import { WarningsCard } from "./components/WarningsCard";
 import {
-  closeStageSession,
-  extractGeometrySession,
   formatUsdErrorForDisplay,
   isInvalidVariantSelectionError,
-  loadPayload,
-  openStageSession,
   parseUsdError,
-  unloadPayload,
   type AssetIssue,
-  type ExtractGeometryOptions,
   type PurposeModes,
   type StageLoadPolicy,
-  type StageSessionHandle,
   type VariantSelection,
 } from "./lib/usd";
 import {
@@ -128,6 +120,7 @@ import {
 import { usePerformanceTracker } from "./hooks/usePerformanceTracker";
 import { useUpdater } from "./hooks/useUpdater";
 import { useUsdInspector } from "./hooks/useUsdInspector";
+import { usePayloadSession } from "./hooks/usePayloadSession";
 
 type SidebarTab = SidebarTabId;
 
@@ -437,33 +430,6 @@ export function App() {
     string | null
   >(null);
 
-  // ---- #44 per-prim payload session ----------------------------------------
-  // When the user opens a USD file with `noPayloads` policy, we also open a
-  // stateful backend session so individual payload prims can be loaded and
-  // unloaded on demand.  The session is closed when the file changes or the
-  // component unmounts.
-  const [stageSessionHandle, setStageSessionHandle] =
-    useState<StageSessionHandle | null>(null);
-  // All SdfPaths that author a payload arc on the current stage. Used to
-  // gate the load/unload buttons in HierarchyCard so they only appear on
-  // genuine payload sources, not on every regular Xform / Mesh.
-  const [payloadPrimPaths, setPayloadPrimPaths] = useState<ReadonlySet<string>>(
-    new Set(),
-  );
-  // Set of SdfPaths whose payload arcs are currently deferred. Derived from
-  // `stageInspection.payloads` + individual load/unload operations.
-  const [unloadedPayloadPaths, setUnloadedPayloadPaths] = useState<
-    ReadonlySet<string>
-  >(new Set());
-  // GLB buffer produced by `extractGeometrySession` after a load/unload.
-  // When non-null, `AssetViewport` should use this buffer instead of
-  // re-extracting from disk. Reset to null on file change AND on variant /
-  // purpose changes (the cached buffer was built against a specific variant
-  // / purpose set; reusing it would freeze the viewport on the snapshot).
-  const [sessionGlbBuffer, setSessionGlbBuffer] = useState<ArrayBuffer | null>(
-    null,
-  );
-
   const isTauri = isTauriEnvironment();
   const debugPanelsEnabled = !isTauri && isDebugPanelsRequested();
   const sidebarCurrentFile = debugPanelsEnabled ? debugPanelFile : currentFile;
@@ -572,12 +538,6 @@ export function App() {
     () => splitViewerWarnings(viewerFeedback.warning),
     [viewerFeedback.warning],
   );
-  const resetUsdCrossCuttingState = useCallback(() => {
-    setVariantSelections([]);
-    setVariantSelectionError(null);
-    setSessionGlbBuffer(null);
-  }, []);
-
   const {
     usdSummary,
     usdInspection,
@@ -585,12 +545,7 @@ export function App() {
     usdLights,
     usdInspectorLoading,
     usdInspectorError,
-  } = useUsdInspector(
-    currentFile,
-    isTauri,
-    usdLoadPolicy,
-    resetUsdCrossCuttingState,
-  );
+  } = useUsdInspector(currentFile, isTauri, usdLoadPolicy);
 
   const warnings = useMemo(() => {
     const nextWarnings: string[] = [];
@@ -717,139 +672,29 @@ export function App() {
     };
   }, []);
 
-  // #44: open a stateful stage session when a USD file is loaded with
-  // `noPayloads` policy (enables per-prim load/unload). Close any previous
-  // session first. When `loadAll` is active no session is needed.
-  useEffect(() => {
-    if (!isTauri || !isUsdFile(currentFile) || !currentFile) {
-      setStageSessionHandle(null);
-      setUnloadedPayloadPaths(new Set());
-      return;
-    }
-
-    // Only open a session when using noPayloads — loadAll doesn't need it.
-    if (usdLoadPolicy !== "noPayloads") {
-      setStageSessionHandle(null);
-      setUnloadedPayloadPaths(new Set());
-      return;
-    }
-
-    let cancelled = false;
-    const path = currentFile.path;
-
-    openStageSession(path, "noPayloads")
-      .then((handle) => {
-        if (cancelled) {
-          // Cleanup ran before this promise resolved — don't leak the
-          // handle on the backend. Issue close in the background and
-          // ignore errors (the registry tolerates missing handles).
-          closeStageSession(handle).catch(() => {});
-          return;
-        }
-        setStageSessionHandle(handle);
-        // Initially all payload sources are unloaded — the session was opened
-        // with noPayloads. The exact set of paths will be populated once
-        // stageInspection settles (via the effect below that syncs payloads).
-      })
-      .catch((err: unknown) => {
-        if (cancelled) return;
-        console.warn(
-          "[usd] open_stage_session failed (per-prim load/unload unavailable):",
-          err,
-        );
-        setStageSessionHandle(null);
-      });
-
-    return () => {
-      cancelled = true;
-      // Close the session asynchronously — we don't await here to avoid
-      // blocking the cleanup. The backend will free the stage.
-      setStageSessionHandle((prev) => {
-        if (prev !== null) {
-          void closeStageSession(prev).catch(() => {
-            // Silently ignore close errors — the Tauri process is likely
-            // already shutting down or the file was closed.
-          });
-        }
-        return null;
-      });
-    };
-  }, [currentFile, isTauri, usdLoadPolicy]);
-
-  // #44: re-sync the session GLB cache when the user changes variants —
-  // but ONLY after the user has actually mutated payloads (i.e. an
-  // override is already in flight). For an untouched session the regular
-  // `requiresGlbPreview`/`extractGeometry` path in loaders.ts handles
-  // variants correctly; forcing a session re-extract here would bypass
-  // that path even on self-contained USDA files that should go through
-  // the Three.js USDLoader.
-  //
-  // Purpose toggles are deliberately NOT a dependency: AssetViewport
-  // applies purpose visibility client-side via `applyPurposeVisibility`,
-  // so the GLB does not need re-extraction when only purpose changes.
-  const sessionGlbBufferRef = useRef<ArrayBuffer | null>(sessionGlbBuffer);
-  useEffect(() => {
-    sessionGlbBufferRef.current = sessionGlbBuffer;
-  }, [sessionGlbBuffer]);
+  const {
+    stageSessionHandle,
+    payloadPrimPaths,
+    unloadedPayloadPaths,
+    sessionGlbBuffer,
+    setSessionGlbBuffer,
+    handleLoadPayload,
+    handleUnloadPayload,
+  } = usePayloadSession(
+    currentFile,
+    isTauri,
+    usdLoadPolicy,
+    usdInspection,
+    variantSelections,
+    purposeModes,
+    recordVariantSelectionError,
+  );
 
   useEffect(() => {
-    if (stageSessionHandle === null) {
-      // No session: drop any leftover override; the stateless extract
-      // path picks up the latest variants on the next render.
-      setSessionGlbBuffer(null);
-      return;
-    }
-    if (sessionGlbBufferRef.current === null) {
-      // Session is open but the user has not yet load/unload-ed any
-      // payload. Leave the override null so loaders.ts uses the regular
-      // `requiresGlbPreview` decision tree.
-      return;
-    }
-    let cancelled = false;
-    extractGeometrySession(stageSessionHandle, {
-      policy: "noPayloads",
-      variantSelections,
-      purposeModes,
-    })
-      .then((buf) => {
-        if (!cancelled) setSessionGlbBuffer(buf);
-      })
-      .catch((err: unknown) => {
-        if (cancelled) return;
-        console.warn("[usd] session re-extract on variant change failed:", err);
-        recordVariantSelectionError(err);
-        setSessionGlbBuffer(null);
-      });
-    return () => {
-      cancelled = true;
-    };
-    // `purposeModes` is captured by closure for defensive completeness
-    // but is intentionally NOT in the deps array — see comment above.
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [variantSelections, stageSessionHandle]);
-
-  // #44: sync payload prim sets from `usdInspection`. `payloadPrimPaths`
-  // contains every prim that authors a payload arc (used to gate the
-  // HierarchyCard load/unload buttons). `unloadedPayloadPaths` is the
-  // currently-deferred subset; later individual load/unload operations
-  // mutate it directly.
-  useEffect(() => {
-    if (!usdInspection || usdLoadPolicy !== "noPayloads") {
-      setPayloadPrimPaths(new Set());
-      setUnloadedPayloadPaths(new Set());
-      return;
-    }
-    const allPayloads = new Set(
-      usdInspection.payloads.map((arc) => arc.sourcePrim),
-    );
-    const unloaded = new Set(
-      usdInspection.payloads
-        .filter((arc) => arc.state === "unloaded")
-        .map((arc) => arc.sourcePrim),
-    );
-    setPayloadPrimPaths(allPayloads);
-    setUnloadedPayloadPaths(unloaded);
-  }, [usdInspection, usdLoadPolicy]);
+    setVariantSelections([]);
+    setVariantSelectionError(null);
+    setSessionGlbBuffer(null);
+  }, [currentFile, usdLoadPolicy, setSessionGlbBuffer]);
 
   const { performanceSnapshot, recordLoadTiming } = usePerformanceTracker(
     settingsPayload,
@@ -1349,6 +1194,7 @@ export function App() {
     };
   }, [canNavigateNext, canNavigatePrev, directoryListing]);
 
+  /* eslint-disable react-hooks/immutability -- useEffectEvent wrappers declared later create stable handles */
   useEffect(() => {
     const handleViewerShortcutDown = (event: KeyboardEvent) => {
       if (event.defaultPrevented) {
@@ -1373,6 +1219,7 @@ export function App() {
       window.removeEventListener("keydown", handleViewerShortcutDown);
     };
   }, []);
+  /* eslint-enable react-hooks/immutability */
 
   const handleOpenFile = async () => {
     try {
@@ -1632,98 +1479,6 @@ export function App() {
       await refreshUpdateConfiguration();
     } catch (error: unknown) {
       setUpdateError(errorMessage(error, "Failed to save updater settings."));
-    }
-  };
-
-  // #44: per-prim payload load/unload. Calls the backend, updates the local
-  // unloaded set optimistically, then re-extracts the GLB from the session.
-  // The re-extract carries the user's variantSelections / purposeModes so a
-  // payload toggle does not silently revert any non-default variant or
-  // purpose mode the user picked from the inspector.
-  //
-  // A ref mirrors the latest `stageSessionHandle` so the async handlers
-  // can detect when the user has navigated away (file change, policy
-  // toggle, app close) between the IPC dispatch and its resolution. In
-  // that case we drop the stale write rather than overwrite the new
-  // file's `sessionGlbBuffer` / `unloadedPayloadPaths` with values from
-  // a session that no longer exists.
-  const stageSessionHandleRef = useRef<StageSessionHandle | null>(
-    stageSessionHandle,
-  );
-  useEffect(() => {
-    stageSessionHandleRef.current = stageSessionHandle;
-  }, [stageSessionHandle]);
-
-  const buildSessionExtractOptions = (): ExtractGeometryOptions => ({
-    policy: "noPayloads",
-    variantSelections,
-    purposeModes,
-  });
-
-  const handleLoadPayload = async (primPath: string) => {
-    const captured = stageSessionHandle;
-    if (captured === null) return;
-    try {
-      await loadPayload(captured, primPath);
-      if (stageSessionHandleRef.current !== captured) return;
-      setUnloadedPayloadPaths((prev) => {
-        const next = new Set(prev);
-        next.delete(primPath);
-        return next;
-      });
-    } catch (err: unknown) {
-      console.error("[usd] load_payload failed:", err);
-      return;
-    }
-    // Re-extract is best-effort: if it fails (e.g. backend hits "no
-    // renderable Mesh prims"), drop the override so the viewport doesn't
-    // keep the pre-load geometry on screen and falls back to the
-    // stateless extract path.
-    try {
-      const glbBuffer = await extractGeometrySession(
-        captured,
-        buildSessionExtractOptions(),
-      );
-      if (stageSessionHandleRef.current !== captured) return;
-      setSessionGlbBuffer(glbBuffer);
-    } catch (err: unknown) {
-      if (stageSessionHandleRef.current !== captured) return;
-      console.warn("[usd] session re-extract after load failed:", err);
-      recordVariantSelectionError(err);
-      setSessionGlbBuffer(null);
-    }
-  };
-
-  const handleUnloadPayload = async (primPath: string) => {
-    const captured = stageSessionHandle;
-    if (captured === null) return;
-    try {
-      await unloadPayload(captured, primPath);
-      if (stageSessionHandleRef.current !== captured) return;
-      setUnloadedPayloadPaths((prev) => {
-        const next = new Set(prev);
-        next.add(primPath);
-        return next;
-      });
-    } catch (err: unknown) {
-      console.error("[usd] unload_payload failed:", err);
-      return;
-    }
-    // Same best-effort re-extract: if the unloaded stage has no
-    // renderable meshes the extract returns an error; clear the override
-    // so the viewport doesn't keep showing the pre-unload geometry.
-    try {
-      const glbBuffer = await extractGeometrySession(
-        captured,
-        buildSessionExtractOptions(),
-      );
-      if (stageSessionHandleRef.current !== captured) return;
-      setSessionGlbBuffer(glbBuffer);
-    } catch (err: unknown) {
-      if (stageSessionHandleRef.current !== captured) return;
-      console.warn("[usd] session re-extract after unload failed:", err);
-      recordVariantSelectionError(err);
-      setSessionGlbBuffer(null);
     }
   };
